Move the gallery/address info into the 'data' folder
