+++ common/views/components/FindUs/FindUs.tsx
@@ -3,7 +3,7 @@ import { font, classNames } from '../../../utils/classnames';
 import {
   wellcomeCollectionGallery,
   wellcomeCollectionAddress,
-} from '../../../model/organization';
+} from '../../../data/organization';
 import Space from '../styled/Space';
 import { FunctionComponent } from 'react';
 import { prismicPageIds } from '../../../services/prismic/hardcoded-id';
